@@ -1,8 +1,8 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :message, :class => PointlessFeedback::Message do
-    name          'A Developer'
-    email_address 'developer@pointlesscorp.com'
-    description   'Site is broke'
-    topic         'Other'
+    name          { 'A Developer' }
+    email_address { 'developer@pointlesscorp.com' }
+    description   { 'Site is broke' }
+    topic         { 'Other' }
   end
 end
