@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,13 +10,13 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20130501182659) do
+ActiveRecord::Schema.define(version: 2013_05_01_182659) do
 
-  create_table "pointless_feedback_messages", force: true do |t|
-    t.string   "name"
-    t.string   "email_address"
-    t.string   "topic"
-    t.text     "description"
+  create_table "pointless_feedback_messages", force: :cascade do |t|
+    t.string "name"
+    t.string "email_address"
+    t.string "topic"
+    t.text "description"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
