@@ -4,7 +4,7 @@ ENV["RAILS_ENV"] = "test"
 require File.expand_path("../dummy/config/environment.rb",  __FILE__)
 require "rails/test_help"
 require "capybara/rails"
-require "factory_girl"
+require "factory_bot"
 require "test/unit"
 require "mocha/setup"
 
